Fix: False faults in fo wires when genfaillog
While inserting fault, if it's a GI fault, genfaillog would insert
the fault at fanout-stem not fanout-branch, which is a false behavior.
This should fix the issue.
Also, the original get_faulty_wire would have alarm message for
GI faults in NOT/BUF gates since the ATPG doesn't generate them.
But since the fault list of genfaillog is from user input(external),
this may happen since the user may insert the fault at that position.
I add a argument to mark if the fault is from external,
and the alarm message would be disabled if external flag is true.